Fix written osu!.db files not being readable by the game
I'm not sure if AccountRank has always been 4 bytes, but it's not trivial to check.

Reported in #18

+++ osu-database-reader/BinaryFiles/OsuDb.cs
@@ -61,6 +61,7 @@ namespace osu_database_reader.BinaryFiles
                 }
             }
 
+            // NOTE: this may be an int instead?
             AccountRank = (PlayerRank)r.ReadByte();
         }
 
@@ -96,7 +97,8 @@ namespace osu_database_reader.BinaryFiles
                 }
             }
 
-            w.Write((byte)AccountRank);
+            // TODO: figure out if/when this changed from byte to int
+            w.Write((int)AccountRank);
         }
     }
 }
